requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
attrs==24.2.0
boto3==1.35.14
botocore==1.35.14
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
datasets==2.21.0
dill==0.3.8
effdet==0.4.1
elastic-transport==8.15.0
elasticsearch==8.15.1
fastapi==0.114.0
filelock==3.16.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
huggingface-hub==0.24.6
idna==3.8
iniconfig==2.0.0
iopath==0.1.10
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
layoutparser==0.3.4
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numpy==2.1.1
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pluggy==1.5.0
portalocker==2.10.1
pyarrow==17.0.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
pyparsing==3.1.4
pypdfium2==4.30.0
pysbd==0.3.4
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
sentencepiece==0.2.0
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
sympy==1.13.2
threadpoolctl==3.5.0
timm==1.0.9
tokenizers==0.19.1
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
-e git+https://github.com/allenai/vila.git@f482972b1d8e1377afc866494f488de81b1d71e5#egg=vila
xxhash==3.5.0
yarl==1.10.0
